server/swarm_analyst/pyproject.toml
[project]
name = "swarm-analyst"
version = "0.1.0"
description = "Parallel multi-agent analyst engine using Gemini 1.5 Pro"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "google-genai>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.4.0",
    "yfinance>=1.2.0",
    "websockets>=13.0",
]
requires-python = ">=3.11"

[tool.uv]
package = true
